# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LeanMate API
  version: 0.1.0
  description: LeanMate V1.1 MVP 接口契约，覆盖认证、用户档案、首页统计、饮食记录、体重记录、AI 日报和连续打卡。
servers:
  - url: https://api.leanmate.app
    description: Production
  - url: https://dev-api.leanmate.app
    description: Development
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /v1/diet/entries:
    get:
      tags: [Diet]
      summary: 查询饮食记录
      description: 返回指定业务日期下未删除的 confirmed 饮食记录。
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
          description: 业务日期。
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFoodEntryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Diet]
      summary: 保存饮食记录
      description: 手动记录可直接保存；AI 草稿需用户确认后保存为正式记录。保存成功后返回记录和刷新后的当日统计快照。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveFoodEntryRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFoodEntrySaveResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/diet/entries/{entryId}:
    put:
      tags: [Diet]
      summary: 编辑饮食记录
      description: 编辑 confirmed 记录后重新计算营养汇总，并刷新对应日期统计快照。
      parameters:
        - $ref: "#/components/parameters/EntryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveFoodEntryRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiFoodEntrySaveResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Diet]
      summary: 删除饮食记录
      description: V1.1 使用软删除，并刷新当日统计快照。
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEmptyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EntryId:
      in: path
      name: entryId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: 40001
            message: 参数错误
            data: null
    Unauthorized:
      description: 未登录或 token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: 40101
            message: 未登录或登录已过期
            data: null
    Forbidden:
      description: 无权限访问资源
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: 40301
            message: 无权限访问该资源
            data: null
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: 40401
            message: 资源不存在
            data: null
...
  schemas:
    ApiErrorResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          nullable: true
    ApiEmptyResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
...
    ApiFoodEntryListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/FoodEntry"
    ApiFoodEntrySaveResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/FoodEntrySaveResult"
...
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
...
    SaveFoodEntryRequest:
      type: object
      required: [mealDate, mealType, sourceType, items]
      properties:
        recognitionTaskId:
          type: string
          format: uuid
          nullable: true
        mealDate:
          type: string
          format: date
        mealType:
          $ref: "#/components/schemas/MealType"
        sourceType:
          type: string
          enum: [photo, text, manual]
        rawText:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SaveFoodItemRequest"
    SaveFoodItemRequest:
      type: object
      required: [name]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          minLength: 1
          maxLength: 128
        quantityText:
          type: string
          maxLength: 128
          nullable: true
        weightG:
          type: number
          nullable: true
        caloriesKcal:
          type: integer
          nullable: true
        proteinG:
          type: number
          nullable: true
        fatG:
          type: number
          nullable: true
        carbsG:
          type: number
          nullable: true
        confidence:
          type: number
          nullable: true
        userEdited:
          type: boolean
          default: false
...
    FoodEntry:
      type: object
      required:
        - id
        - mealDate
        - mealType
        - sourceType
        - status
        - totalCaloriesKcal
        - totalProteinG
        - totalFatG
        - totalCarbsG
        - items
      properties:
        id:
          type: string
          format: uuid
        recognitionTaskId:
          type: string
          format: uuid
          nullable: true
        mealDate:
          type: string
          format: date
        mealType:
          $ref: "#/components/schemas/MealType"
        sourceType:
          type: string
          enum: [photo, text, manual]
        rawText:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, confirmed, deleted]
        totalCaloriesKcal:
          type: integer
        totalProteinG:
          type: number
        totalFatG:
          type: number
        totalCarbsG:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/FoodItem"
        createdAt:
          type: string
          format: date-time
...
    FoodEntrySaveResult:
      type: object
      required: [entry, today]
      properties:
        entry:
          $ref: "#/components/schemas/FoodEntry"
        today:
          $ref: "#/components/schemas/DailyNutritionSnapshot"
    FoodItem:
      type: object
      required: [id, name, userEdited]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        quantityText:
          type: string
          nullable: true
        weightG:
          type: number
          nullable: true
        caloriesKcal:
          type: integer
          nullable: true
        proteinG:
          type: number
          nullable: true
        fatG:
          type: number
          nullable: true
        carbsG:
          type: number
          nullable: true
        confidence:
          type: number
          nullable: true
        userEdited:
          type: boolean
...
    DailyNutritionSnapshot:
      type: object
      required:
        - date
        - calorieTargetKcal
        - caloriesKcal
        - remainingCaloriesKcal
        - proteinG
        - fatG
        - carbsG
        - foodEntryCount
      properties:
        date:
          type: string
          format: date
        calorieTargetKcal:
          type: integer
        caloriesKcal:
          type: integer
        remainingCaloriesKcal:
          type: integer
        proteinG:
          type: number
        fatG:
          type: number
        carbsG:
          type: number
        foodEntryCount:
          type: integer
        weightKg:
          type: number
          nullable: true
...
    MealType:
      type: string
      enum: [breakfast, lunch, dinner, snack]